==> 1. anomaly detection ETL.ipynb <==
Import needed libraries

import requests
import pandas as pd
import numpy as np
import csv
import io
import datetime
import pyodbc
import urllib.request
import json
import pyodbc
import sqlalchemy
from pandas.io.json import json_normalize
import time

ETL raw data from government API at https://data.go.th/ and save as a pickle file

years = {
        '2558' : 'lurONURJJ4a2kCYMt41ZTBuqOCUQvMmp', #used today 
        '2559' : 'aQ92erVTgdvIE7g9zivJGTwOllp0exmR', #used yesterday
        '2560' : 'oAA5nVzZQ6iLTc77L9FXaAyfbS3Z4dpJ', # used for 2559
        '2561' : 'EM5bdWNNsdp7MND4URk76KEWS49yGCMB',
        }
  
#  LMnzeNfoOPRk79eKVpX3e1IWoVtJ2Pfl used
#  NveGgxnickjpTodTKtlo9j319fOHt2eC 


project_list = []
contract_list = []
url = ("https://opend.data.go.th/govspending/cgdcontract?api-key=NveGgxnickjpTodTKtlo9j319fOHt2eC&year=2562&"
       + "budget_start=1000000&limit=100&offset=")
for i in range(0,5):
    start = time.time()
    offset = str(100*i)
    print("getting row number " + offset)
    for trial in [1,2,3,4,5]:
        try:
            response = urllib.request.urlopen(url+offset)
            html = response.read()
            break;
        except:
            print("Something is wrong! Retry at row " + offset + " try again! for the " + str(trial) + " times out of 5 times")

    json_obj = json.loads(html)
    df_temp = pd.DataFrame(json_obj['result'])
    df_contract_temp = pd.json_normalize(data=json_obj['result'], record_path='contract', meta = ['project_id'])
    project_list.append(df_temp)
    contract_list.append(df_contract_temp)
    end = time.time()
    print(str(i) + ' trial takes ' + str(end - start)+ ' seconds')

df_project = pd.concat(project_list)
df_contract = pd.concat(contract_list)

df_project.project_id.nunique()

print(df_project.shape)
print(df_contract.shape)

df_project.head()

df_project.to_pickle('Raw_project.pkl')
df_contract.to_pickle('Raw_contract.pkl')


project_type_name_th = df.project_type_name.unique()
project_type_name_th = ['จ้างทำของ/จ้างเหมาบริการ', 'จ้างก่อสร้าง', 'เช่า', 'ซื้อ','จ้างที่ปรึกษา']
project_type_name_eng = ['Contracting services / Contracting services', 'Hire for construction',
                         'hire','buy','Hire a consultant']
project_type_name_dict = dict(zip(project_type_name_th,project_type_name_eng))


dept_name_th = ['กองทัพเรือ', 'กองทัพอากาศ', 'การไฟฟ้าส่วนภูมิภาค',
       'การทางพิเศษแห่งประเทศไทย', 'การไฟฟ้าฝ่ายผลิตแห่งประเทศไทย',
       'สำนักงานตำรวจแห่งชาติ', 'กรมทางหลวง',
       'สำนักงานปลัดกระทรวงการคลัง', 'กรมชลประทาน',
       'การรถไฟฟ้าขนส่งมวลชนแห่งประเทศไทย', 'สถาบันการบินพลเรือน',
       'กรมการขนส่งทางบก', 'กรมท่าอากาศยาน', 'การประปาส่วนภูมิภาค',
       'กรมฝนหลวงและการบินเกษตร', 'สำนักงานหลักประกันสุขภาพแห่งชาติ',
       'สำนักงานปลัดกระทรวงกลาโหม', 'กองบัญชาการกองทัพไทย',
       'สถาบันเทคโนโลยีนิวเคลียร์แห่งชาติ (องค์การมหาชน)',
       'กรมประชาสัมพันธ์', 'กรมบัญชีกลาง']

dept_name_eng = ['Navy', 'Air Force', 'Provincial Electricity Authority',
    'Expressway Authority of Thailand', 'Electricity Generating Authority of Thailand',
    'National Police Agency', 'Department of Highways',
    'Office of the Permanent Secretary for Finance', 'Irrigation Department',
    'Mass Rapid Transit Authority of Thailand', 'Civil Aviation Training Center',
    'Department of Land Transport', 'Department of Airport', 'Provincial Waterworks Authority',
    'Department of Royal Rainmaking and Agricultural Aviation', 'National Health Security Office',
    'Office of the Permanent Secretary for Defense', 'Thai Armed Forces Headquarters',
    'National Institute of Nuclear Technology (Public organization) ',
    'Public Relations Department', 'Comptroller Generals Department']




df_project_temp = df_project.replace(project_type_name_th,['Contracting services / Contracting services', 'Hire for construction',
                         'hire','buy','Hire a consultant'])

df_project_temp = pd.read_pickle('2558_project.pkl')
