==> 3. anomaly detection Data Cleaning.ipynb <==
import requests
import pandas as pd
import numpy as np
import csv
import io
import datetime
from sqlalchemy import create_engine
import pyodbc
import urllib.request
import json
import pyodbc
from sqlalchemy import create_engine
import sqlalchemy
from pandas.io.json import json_normalize
import time
import dateparser
from googletrans import Translator
import timeit
import codecs
import pickle
import pythainlp

Clean data and make sure they are in the correct format

# open pickle of date
with open("unique_date_dict.pkl", 'rb') as f:
    temp = pickle.load(f)

#put it as DataFrame object
df_temp = pd.Series(temp).to_frame()
df_temp_2 = df_temp.reset_index()
df_temp_2

#create temp value
df_temp_2["temp_eng"] =  df_temp_2[0]

#break date value into it own col
df_temp_2['day'], df_temp_2['month'], df_temp_2['year'] = df_temp_2['temp_eng'].str.split(' ', 2).str

#replace &#39; with ''
df_temp_2['year'].replace('&#39;', '', regex=True,inplace = True)


#if 2 digits add 20 to the front

def add_20(x):
    if isinstance(x , str):
        if ( len(x) <= 2):
            x = '20' + x
    return x

df_temp_2['year'].unique()

df_temp_2['year'] = df_temp_2['year'].apply(add_20)

df_temp_2.year.unique()

def change_year(x):
    if isinstance(x , str):
        if ( int(x) > 2050):
            x = int(x) - 43
    return str(x)

df_temp_2['year'] = df_temp_2['year'].apply(change_year)

df_temp_2

df_temp_2['Eng_date_cleaned'] = df_temp_2['day'] + ' ' + df_temp_2['month'] + ' ' + df_temp_2['year']

df_temp_2

df_temp_3 = df_temp_2[['index','Eng_date_cleaned']]

unique_date_dict_cleaned = {}
unique_date_dict_key = df_temp_3['index']
unique_date_dict_value = df_temp_3['Eng_date_cleaned']

res = dict(zip(unique_date_dict_key, unique_date_dict_value))

f = open("unique_date_dict_cleaned.pkl","wb")
pickle.dump(res,f)
f.close()

df_contract = pd.read_csv("contract_eng_consolidated.csv")
df_project = pd.read_csv("project_eng_consolidated.csv")

df_project.head()
