==> 4. Anomaly detection Feature Engineering.ipynb <==
import requests
import pandas as pd
import numpy as np
import csv
import io
import datetime
from sqlalchemy import create_engine
import pyodbc
import urllib.request
import json
import pyodbc
from sqlalchemy import create_engine
import sqlalchemy
from pandas.io.json import json_normalize
import time
import dateparser
from googletrans import Translator
import timeit
import codecs
import pickle
import pythainlp

from pythainlp import sent_tokenize, word_tokenize
from pythainlp.corpus import wordnet
from nltk.stem.porter import PorterStemmer
from nltk.corpus import words
from nltk.corpus import stopwords
from gensim import corpora
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import LatentDirichletAllocation as LDA
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.decomposition import TruncatedSVD
from sklearn.cluster import KMeans
from sklearn.decomposition import NMF

1. create dept grouping manually using IF contain text statement

2. Create topic grouping for project name
    2.1: Tokenize, remove stop word, remove unwanted word from project name
    2.2: Use tf idf to assign score for each word
    2.3: Use NMF to create 20 topic grouping.

df_contract = pd.read_csv("contract_eng_consolidated.csv")
df_project = pd.read_csv("project_eng_consolidated.csv")

df_project.shape



def dept_name_grouping(x):
    x = x.lower()
    if "hospital" in x:
        return "Hospital"
    elif any(ele in x for ele in ['university','campus','institute','college','faculty']):
        return "Higher education"
    elif any(ele in x for ele in ['school','kindergarten']):
        return 'School'
    elif any(ele in x for ele in ['municipality','subdistrict','sub-district','district','provincial']):
        return 'Local Government unit'
    elif any(ele in x for ele in ['office','department','government']):
        return 'department'
    elif 'provincial' in x:
        return 'provincial'
    elif 'police' in x:
        return 'police'
    return 'Other'

#df_project['dept_name_group'] = df_project['dept_name_eng'].apply(dept_name_grouping)

#df_project.to_csv("project_eng_consolidated.csv", index = False)

Topic_grouping = df_project[['project_name','dept_name_eng']]

#### Topic Modeling ########

## replace thai num with english num
# thai number in the list
num_map = {}
num_eng = 0
for num_thai in pythainlp.thai_digits:
    num_map[str(num_thai)] = str(num_eng)
    num_eng += 1
Topic_grouping['project_name'].replace(num_map, inplace=True, regex=True)

## function to remove stop word and unwanted word

## get word that have length greater than....
## get word that have count greater than...
def preprocess_data(txt_string):
    tokens = word_tokenize(txt_string, keep_whitespace=False)
    tokens = [i for i in tokens if not i in th_stop and not i in en_stop and not i in th_manual_word and len(i) > 2]
    return tokens

# put tokenize word back into list and join them using ,
def tokenize_text(raw_col):
    tokens_list = [preprocess_data(txt) for txt in raw_col]
    tokens_list_j = [','.join(tkn) for tkn in tokens_list]
    return tokens_list_j

en_stop = set(stopwords.words('english'))
th_stop = pythainlp.corpus.thai_stopwords()
th_manual_word = ['ประกวดราคา', 'จ้าง','ทางอิเล็กทรอนิกส์','ซื้อ','กรณีพิเศษ','(',')' ,'/' ,'+','-' ,'e-bidding',
                  'อิเล็กทรอนิกส์']

project_type_name_th = ['จ้างก่อสร้าง', 'จ้างทำของ/จ้างเหมาบริการ', 'ซื้อ', 'เช่า',
       'จ้างที่ปรึกษา', 'จ้างควบคุมงาน', 'จ้างออกแบบ',
       'จ้างออกแบบและควบคุมงาน']

purchase_method_group_name_th = ['ประกวดราคา', 'ประกวดราคาด้วยวิธีการทางอิเล็กทรอนิกส์', 'พิเศษ',
       'จ้างที่ปรึกษา', 'กรณีพิเศษ',
       'ประกวดราคาอิเล็กทรอนิกส์ (e-bidding)', 'สอบราคา', 'ตกลงราคา',
       'ตลาดอิเล็กทรอนิกส์ (e-market)', 'จ้างออกแบบ', 'คัดเลือก',
       'เฉพาะเจาะจง']

purchase_method_group_name_eng = ['Bidding', 'E-bidding', 'Special',
        'Hire consultants', 'Special cases',
        'E-bidding','Price examination','Agreed price',
        'E-market', 'Hire to design', 'selection',
        'Specific']

province_th = ['นนทบุรี', 'กรุงเทพมหานคร', 'อุตรดิตถ์', 'เชียงราย', 'ชลบุรี',
       'พระนครศรีอยุธยา', 'สมุทรปราการ', 'ปทุมธานี', 'ลำปาง', 'นครนายก',
       'เชียงใหม่', 'ขอนแก่น', 'พะเยา', 'ปัตตานี', 'ลพบุรี',
       'อุบลราชธานี', 'ยะลา', 'นครราชสีมา', 'อุดรธานี', 'สมุทรสาคร',
       'บึงกาฬ', 'จันทบุรี', 'ฉะเชิงเทรา', 'นครศรีธรรมราช', 'สงขลา',
       'นครปฐม', 'มหาสารคาม', 'กาญจนบุรี', 'แพร่', 'พิจิตร', 'ภูเก็ต',
       'พิษณุโลก', 'กาฬสินธุ์', 'กำแพงเพชร', 'สุราษฎร์ธานี', 'ระยอง',
       'ศรีสะเกษ', 'เพชรบุรี', 'เลย', 'ชัยภูมิ', 'สกลนคร', 'นครพนม',
       'บุรีรัมย์', 'สระบุรี', 'ปราจีนบุรี', 'สุรินทร์', 'นครสวรรค์',
       'เพชรบูรณ์', 'นราธิวาส', 'ประจวบคีรีขันธ์', 'ลำพูน', 'ยโสธร',
       'สุโขทัย', 'สตูล', 'พังงา', 'ชัยนาท', 'กระบี่', 'พัทลุง', 'ชุมพร',
       'ตาก', 'ตรัง', 'อำนาจเจริญ', 'ราชบุรี', 'ร้อยเอ็ด', 'หนองคาย',
       'มุกดาหาร', 'สิงห์บุรี', 'สมุทรสงคราม', 'สุพรรณบุรี', 'อ่างทอง',
       'ตราด', 'ระนอง', 'น่าน', 'อุทัยธานี', 'สระแก้ว', 'หนองบัวลำภู',
       'แม่ฮ่องสอน']

th_manual_word = th_manual_word + project_type_name_th + purchase_method_group_name_th + purchase_method_group_name_eng +province_th

# create a column of tokenize word
Topic_grouping['clean_project_name'] = tokenize_text(Topic_grouping['project_name'])


#Tfidf: Count how often it occur in each document as well as how often it occur overall
# if occur a lot in that document, the word is important, but if occur a lot in overall, not as important
tvec = TfidfVectorizer(analyzer=lambda x:x.split(','),)
X = tvec.fit_transform(Topic_grouping['clean_project_name'])

######## Method 3: LDA & NMF ##########


#helper method to print top 10 topics
def print_top_words(model, feature_names, n_top_words):
    for topic_idx, topic in enumerate(model.components_):
        message = "Topic #%d: " % topic_idx
        message += " ".join([feature_names[i]
                             for i in topic.argsort()[:-n_top_words - 1:-1]])
        print(message)
    print()
#####
tf_feature_names = tvec.get_feature_names()

#### NMF
nmf = NMF(n_components=20, random_state=1,
          beta_loss='kullback-leibler', solver='mu', max_iter=1000, alpha=.1,
          l1_ratio=.5).fit(X)

print_top_words(nmf, tf_feature_names, 50)

Factor A into W,H

Use W to get topics

W = nmf.fit_transform(X)

assign_topic = np.argmax(W,axis = 1)

for i in range(nmf.n_components):
    print("Number of documents in topics ", i," is ", np.sum(assign_topic==i), " and ", np.sum(assign_topic==i)/len(assign_topic) *100, '% of total ')

### find location where 
### 
'''
feature_names[i]
                             for i in topic.argsort()[:-n_top_words - 1:-1]])
                             
1. return index of the array in order to sort it from 0 to max
2. 

    '''


df_project["Topic_group"] = assign_topic

#df_project.to_csv('project_eng_consolidated_with_Topics.csv',index=False)

df_project.head()
